[Refactor] refactor multiview humanpose (#1196)
* 'multiview_pose'

* 'single_pose'

* multiview_pose_detect_and_regress

* unit test

* typo

* typo

* unit test

* del old codes

* del old codes

* 'fix_test_bug'

* 'accuray_test'

* fix input_heatmaps

* stay consistent with official voxelpose model keys

* fix lint

* add comments
