+++ mmpose/models/detectors/__init__.py
@@ -3,12 +3,14 @@ from .associative_embedding import AssociativeEmbedding
 from .interhand_3d import Interhand3D
 from .mesh import ParametricMesh
 from .multi_task import MultiTask
+from .multiview_pose import (DetectAndRegress, VoxelCenterDetector,
+                             VoxelSinglePose)
 from .pose_lifter import PoseLifter
 from .posewarper import PoseWarper
 from .top_down import TopDown
-from .voxelpose import VoxelPose
 
 __all__ = [
     'TopDown', 'AssociativeEmbedding', 'ParametricMesh', 'MultiTask',
-    'PoseLifter', 'Interhand3D', 'PoseWarper', 'VoxelPose'
+    'PoseLifter', 'Interhand3D', 'PoseWarper', 'DetectAndRegress',
+    'VoxelCenterDetector', 'VoxelSinglePose'
 ]
